Prerunchecker, change mediaobject, filecontroller
Prerunchecker checks what mediaobjects that don't have a file in their 
path after loading a session. Gives the user a popup box with these

Change mediaobject should be done for the modules. Needs to be called
through the GUI. (Need modal for GUI option for changing mediaObject
path)

FileController now has control over which filetypes we accept, and it
has methods to get fileextension and MediaSourceType from a filepath